--- ppcls/arch/gears/contrastivehead.py
@@ -1,53 +1,0 @@
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import paddle
-import paddle.nn as nn
-
-
-class ContrastiveHead(nn.Layer):
-    """Head for contrastive learning.
-
-    Args:
-        temperature (float): The temperature hyper-parameter that
-            controls the concentration level of the distribution.
-            Default: 0.1.
-    """
-
-    def __init__(self, temperature=0.1, **kwargs):
-        super(ContrastiveHead, self).__init__()
-        self.criterion = nn.CrossEntropyLoss()
-        self.temperature = temperature
-
-    def forward(self, pos, neg):
-        """Forward head.
-
-        Args:
-            pos (Tensor): Nx1 positive similarity.
-            neg (Tensor): Nxk negative similarity.
-
-        Returns:
-            dict[str, Tensor]: A dictionary of loss components.
-        """
-        N = pos.shape[0]
-        logits = paddle.concat((pos, neg), axis=1)
-        logits /= self.temperature
-        labels = paddle.zeros((N, ), dtype='int64')
-        outputs = dict()
-        outputs['loss'] = self.criterion(logits, labels)
-        return outputs
