--- ppcls/arch/gears/moconeck.py
@@ -1,58 +1,0 @@
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserve.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import paddle.nn as nn
-
-
-class LinearNeck(nn.Layer):
-    """Linear neck: fc only.
-    """
-
-    def __init__(self, in_channels, out_channels, with_avg_pool=False):
-        super(LinearNeck, self).__init__()
-        self.with_avg_pool = with_avg_pool
-        if with_avg_pool:
-            self.avgpool = nn.AdaptiveAvgPool2D((1, 1))
-        self.fc = nn.Linear(in_channels, out_channels)
-
-    def forward(self, x):
-
-        if self.with_avg_pool:
-            x = self.avgpool(x)
-        return self.fc(x.reshape([x.shape[0], -1]))
-
-
-class NonLinearNeck(nn.Layer):
-    """The non-linear neck in MoCo v2: fc-relu-fc.
-    """
-
-    def __init__(self,
-                 in_channels,
-                 hid_channels,
-                 out_channels,
-                 with_avg_pool=False):
-        super(NonLinearNeck, self).__init__()
-        self.with_avg_pool = with_avg_pool
-        if with_avg_pool:
-            self.avgpool = nn.AdaptiveAvgPool2D((1, 1))
-
-        self.mlp = nn.Sequential(
-            nn.Linear(in_channels, hid_channels),
-            nn.ReLU(), nn.Linear(hid_channels, out_channels))
-
-    def forward(self, x):
-
-        if self.with_avg_pool:
-            x = self.avgpool(x)
-        return self.mlp(x.reshape([x.shape[0], -1]))
